+++ packages/types/src/subscription.ts
@@ -4,7 +4,8 @@ export type Subscription =
   Flowglad.Subscriptions.SubscriptionRetrieveResponse['subscription']
 
 export type SubscriptionItem =
-  Flowglad.CustomerRetrieveBillingResponse.Subscription.SubscriptionItem
+  | Flowglad.CustomerRetrieveBillingResponse.SubscriptionItem.StaticSubscriptionItem
+  | Flowglad.CustomerRetrieveBillingResponse.SubscriptionItem.UsageSubscriptionItem
 
 export type SubscriptionStatus =
   Flowglad.Subscriptions.SubscriptionRetrieveResponse.Subscription['status']
@@ -14,3 +15,6 @@ export type SubscriptionIntervalUnit =
 
 export type SubscriptionDetails =
   Flowglad.CustomerRetrieveBillingResponse['subscriptions'][number]
+
+export type SubscriptionExperimentalFields =
+  Flowglad.CustomerRetrieveBillingResponse.SubscriptionItem.Experimental