+++ platform/flowglad-next/src/subscriptions/billingPeriodHelpers.ts
@@ -262,6 +262,7 @@ export const attemptToTransitionSubscriptionBillingPeriod = async (
       eventsToLog: [],
     }
   }
+
   const allBillingPeriods = await selectBillingPeriods(
     { subscriptionId: subscription.id },
     transaction