+++ packages/types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @flowglad/types
 
+## 0.10.17
+
+### Patch Changes
+
+- Remove flowglad-root root theming
+
+## 0.10.16
+
+### Patch Changes
+
+- Add getProduct and getPrice to SDK, support activate_subscription checkout sessions
+
 ## 0.10.15
 
 ### Patch Changes